move ato to module folder
